move otp to dedicated package under util
