+++ internal/util/duration/duration.go
